+++ tket/src/OpType/OpTypeInfo.cpp
@@ -101,6 +101,8 @@ const std::map<OpType, OpTypeInfo>& optypeinfo() {
         "PauliExpCommutingSetBox",
         {},
         std::nullopt}},
+      {OpType::TermSequenceBox,
+       {"TermSequenceBox", "TermSequenceBox", {}, std::nullopt}},
       {OpType::CustomGate, {"CustomGate", "CustomGate", {}, std::nullopt}},
       {OpType::Barrier, {"Barrier", "Barrier", {}, std::nullopt}},
       {OpType::Measure,