+++ pytket/docs/changelog.rst
@@ -17,7 +17,10 @@ Features:
   ``BackendResult.get_probability_distribution()`` and to the constructor of a
   ``ProbabilityDistribution``, defaulting to zero. (Previously probabilities
   below 1e-10 were by default treated as zero.)
-
+* Add python binding for ``UnitaryRevTableau``.
+* Add ``TermSequenceBox``, for circuit synthesis of a series of Pauli 
+  Exponentials, where the ordering of terms can be changed.
+  
 Fixes:
 
 * Add missing op types to methods for converting Clifford circuits to unitary